standby: update user_scene_lock & pwr_config
    1. add user_scene_lock;
    2. update super_standby power config;
    3. update sys pwr usage: add vcc-lpddr;
    4. code reconstruct: for pm_debug
    5. bugfix: for sys_pwr_dm_store

+++ include/linux/power/axp_depend.h
@@ -176,6 +176,6 @@ extern int is_sys_pwr_dm_id(const char *id);
 extern int is_sys_pwr_dm_active(unsigned int bitmap);
 extern char *get_sys_pwr_dm_id(unsigned int bitmap);
 extern unsigned int (* get_pwr_regu_tree(void))[VCC_MAX_INDEX];
-extern unsigned int parse_pwr_dm_map(char *s, unsigned int size, unsigned int bitmap);
+extern ssize_t parse_pwr_dm_map(char *s, size_t size, unsigned int bitmap);
 #endif /* __AXP_DEPEND_H__ */
